Added launcher interface and registry
